+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Changelog
 
-## 3.0.5 - 3.0.6
+## 3.0.5 - 3.0.7
 
--- Break components and layout css into separate files (so one can be dequeued without the other if needed)
--- Fix title of global modules
+- Break components and layout css into separate files (so one can be dequeued without the other if needed)
+- Fix title of global modules
+- Add overwrite mode back into the settings
+- Make slider options "perPage" responsive.
 
 ## 3.0.3 - 3.0.4